+++ e2e/__tests__/transform.test.ts
@@ -185,3 +185,25 @@ describe('transformer-config', () => {
     expect(stdout).toMatchSnapshot();
   });
 });
+
+describe('transformer caching', () => {
+  const dir = path.resolve(__dirname, '..', 'transform/cache');
+  const dirRegexp = new RegExp('^' + dir);
+
+  it('does not rerun transform within worker', () => {
+    // --no-cache because babel can cache stuff and result in false green
+    const {stdout} = runJest(dir, ['--no-cache', '-w=2']);
+
+    const loggedFiles = stdout.split('\n').map(line => {
+      expect(line).toMatch(dirRegexp);
+
+      return line.replace(dirRegexp, '<rootDir>');
+    });
+
+    // We run with 2 workers, so the file should be transformed twice
+    expect(loggedFiles).toEqual([
+      '<rootDir>/common-file.js',
+      '<rootDir>/common-file.js',
+    ]);
+  });
+});